requirements.txt
azure-cognitiveservices-language-luis~=0.2.0
botbuilder-integration-aiohttp>=4.13.0
botbuilder-ai>=4.13.0
ibm-watson>=4.7.1
